Moved tests to parent package
