+++ src/test/java/fr/motormingle/api/ManufacturerControllerTest.java
@@ -1,5 +1,6 @@
-package fr.motormingle.api.controller;
+package fr.motormingle.api;
 
+import fr.motormingle.api.controller.ManufacturerController;
 import fr.motormingle.api.entity.Manufacturer;
 import fr.motormingle.api.exception.NotFoundException;
 import fr.motormingle.api.service.ManufacturerService;
