+++ src/test/java/fr/motormingle/api/MotorbikeControllerTest.java
@@ -1,5 +1,6 @@
-package fr.motormingle.api.controller;
+package fr.motormingle.api;
 
+import fr.motormingle.api.controller.MotorbikeController;
 import fr.motormingle.api.entity.Motorbike;
 import fr.motormingle.api.exception.NotFoundException;
 import fr.motormingle.api.service.MotorbikeService;
